+++ tests/models/README.md
@@ -1,0 +1,2 @@
+* All the tests in `models/nm_models_core` run by default
+* The tests in `models` require settng the `TEST_ALL_MODELS=1` enviornment variable
