+++ tests/models/test_big_models.py
@@ -9,6 +9,12 @@ import sys
 
 import pytest
 
+from tests.nm_utils.utils_skip import should_skip_test_group
+
+if should_skip_test_group(group_name="TEST_ALL_MODELS"):
+    pytest.skip("TEST_ALL_MODELS=0, skipping kernel group",
+                allow_module_level=True)
+
 MODELS = [
     "meta-llama/Llama-2-7b-hf",
     "mistralai/Mistral-7B-v0.1",
