+++ tests/models/test_fp8.py
@@ -8,8 +8,13 @@ import pytest
 import torch
 from transformers import AutoTokenizer
 
+from tests.nm_utils.utils_skip import should_skip_test_group
 from vllm import LLM, SamplingParams
 from vllm.model_executor.layers.quantization import QUANTIZATION_METHODS
+
+if should_skip_test_group(group_name="TEST_ALL_MODELS"):
+    pytest.skip("TEST_ALL_MODELS=0, skipping kernel group",
+                allow_module_level=True)
 
 os.environ["TOKENIZERS_PARALLELISM"] = "true"
