+++ tests/models/test_llava.py
@@ -7,7 +7,12 @@ import pytest
 import torch
 from transformers import AutoTokenizer
 
+from tests.nm_utils.utils_skip import should_skip_test_group
 from vllm.config import VisionLanguageConfig
+
+if should_skip_test_group(group_name="TEST_ALL_MODELS"):
+    pytest.skip("TEST_ALL_MODELS=0, skipping kernel group",
+                allow_module_level=True)
 
 model_and_vl_config = [
     ("llava-hf/llava-1.5-7b-hf",
