+++ tests/models/test_models.py
@@ -9,6 +9,12 @@ Run `pytest tests/models/test_models.py`.
 """
 import pytest
 
+from tests.nm_utils.utils_skip import should_skip_test_group
+
+if should_skip_test_group(group_name="TEST_ALL_MODELS"):
+    pytest.skip("TEST_ALL_MODELS=0, skipping kernel group",
+                allow_module_level=True)
+
 MODELS = [
     "facebook/opt-125m",
     "gpt2",
