+++ tests/models/test_registry.py
@@ -1,6 +1,11 @@
 import pytest
 
+from tests.nm_utils.utils_skip import should_skip_test_group
 from vllm.model_executor.models import _MODELS, ModelRegistry
+
+if should_skip_test_group(group_name="TEST_ALL_MODELS"):
+    pytest.skip("TEST_ALL_MODELS=0, skipping kernel group",
+                allow_module_level=True)
 
 
 @pytest.mark.parametrize("model_cls", _MODELS)
